Move some test headers into core/utils
