+++ tests/benchmarks/cuda/benchmark_propagator_cuda.cpp
@@ -13,7 +13,7 @@
 #include <vecmem/memory/cuda/managed_memory_resource.hpp>
 
 #include "benchmark_propagator_cuda_kernel.hpp"
-#include "tests/common/tools/track_generators.hpp"
+#include "detray/simulation/track_generators.hpp"
 #include "vecmem/utils/cuda/copy.hpp"
 
 using namespace detray;
