+++ tests/common/include/tests/common/check_geometry_navigation.inl
@@ -11,17 +11,17 @@
 #include <vecmem/memory/host_memory_resource.hpp>
 
 #include "detray/definitions/units.hpp"
+#include "detray/detectors/create_toy_geometry.hpp"
 #include "detray/intersection/detail/trajectories.hpp"
 #include "detray/propagator/actor_chain.hpp"
 #include "detray/propagator/line_stepper.hpp"
 #include "detray/propagator/navigator.hpp"
 #include "detray/propagator/propagator.hpp"
 #include "detray/propagator/rk_stepper.hpp"
+#include "detray/simulation/track_generators.hpp"
 #include "detray/tracks/tracks.hpp"
-#include "tests/common/tools/create_toy_geometry.hpp"
 #include "tests/common/tools/inspectors.hpp"
 #include "tests/common/tools/particle_gun.hpp"
-#include "tests/common/tools/track_generators.hpp"
 
 using namespace detray;
