+++ tests/common/include/tests/common/check_geometry_scan.inl
@@ -10,13 +10,13 @@
 #include <iostream>
 #include <vecmem/memory/host_memory_resource.hpp>
 
+#include "detray/detectors/create_toy_geometry.hpp"
 #include "detray/geometry/volume_graph.hpp"
 #include "detray/intersection/detail/trajectories.hpp"
-#include "tests/common/tools/create_toy_geometry.hpp"
+#include "detray/simulation/track_generators.hpp"
 #include "tests/common/tools/hash_tree.hpp"
 #include "tests/common/tools/particle_gun.hpp"
 #include "tests/common/tools/ray_scan_utils.hpp"
-#include "tests/common/tools/track_generators.hpp"
 
 /// @note __plugin has to be defined with a preprocessor command
 using namespace detray;
