+++ tests/common/include/tests/common/io_read_detector.inl
@@ -13,7 +13,7 @@
 #include <vecmem/memory/host_memory_resource.hpp>
 
 #include "detray/core/detector.hpp"
-#include "tests/common/tools/detector_metadata.hpp"
+#include "detray/detectors/detector_metadata.hpp"
 #include "tests/common/tools/read_geometry.hpp"
 
 /// @note __plugin has to be defined with a preprocessor command
