+++ tests/common/include/tests/common/line_stepper.inl
@@ -7,13 +7,13 @@
 
 // Project include(s).
 #include "detray/definitions/units.hpp"
+#include "detray/detectors/create_telescope_detector.hpp"
 #include "detray/propagator/actor_chain.hpp"
 #include "detray/propagator/actors/parameter_resetter.hpp"
 #include "detray/propagator/actors/parameter_transporter.hpp"
 #include "detray/propagator/line_stepper.hpp"
 #include "detray/propagator/navigator.hpp"
 #include "detray/propagator/propagator.hpp"
-#include "tests/common/tools/create_telescope_detector.hpp"
 
 // Covfie include(s).
 #include <covfie/core/field.hpp>
