+++ tests/common/include/tests/common/test_toy_geometry.inl
@@ -9,7 +9,7 @@
 #include <type_traits>
 
 #include "detray/definitions/detail/accessor.hpp"
-#include "tests/common/tools/create_toy_geometry.hpp"
+#include "detray/detectors/create_toy_geometry.hpp"
 
 using namespace detray;
