+++ tests/common/include/tests/common/tools_particle_gun.inl
@@ -7,11 +7,11 @@
 
 // Project include(s).
 #include "detray/definitions/units.hpp"
+#include "detray/detectors/create_toy_geometry.hpp"
 #include "detray/intersection/detail/trajectories.hpp"
+#include "detray/simulation/track_generators.hpp"
 #include "detray/tracks/tracks.hpp"
-#include "tests/common/tools/create_toy_geometry.hpp"
 #include "tests/common/tools/particle_gun.hpp"
-#include "tests/common/tools/track_generators.hpp"
 
 // VecMem include(s).
 #include <vecmem/memory/host_memory_resource.hpp>
