+++ tests/unit_tests/cuda/navigator_cuda_kernel.hpp
@@ -21,10 +21,10 @@
 #include <covfie/core/field_view.hpp>
 
 #include "detray/definitions/units.hpp"
+#include "detray/detectors/create_toy_geometry.hpp"
 #include "detray/propagator/line_stepper.hpp"
 #include "detray/propagator/navigator.hpp"
-#include "tests/common/tools/create_toy_geometry.hpp"
-#include "tests/common/tools/track_generators.hpp"
+#include "detray/simulation/track_generators.hpp"
 
 using namespace detray;
