+++ utils/include/detray/detectors/create_toy_geometry.hpp
@@ -11,7 +11,7 @@
 #include "detray/core/detector.hpp"
 #include "detray/definitions/units.hpp"
 #include "detray/materials/predefined_materials.hpp"
-#include "tests/common/tools/detector_metadata.hpp"
+#include "detray/detectors/detector_metadata.hpp"
 
 // Vecmem include(s)
 #include <vecmem/memory/host_memory_resource.hpp>
