+++ utils/include/detray/detectors/detector_metadata.hpp
